+++ Screens/GR_Filter.lua
@@ -18,89 +18,88 @@ end
 function filter:Init()
     self.tblFrame = {}
     self.tblFilter = {}
-    self.activeRecord = nil
+    self.nameOk = false
 
     self.tblRaceByFileName = {}
     self.tblClassByFileName = {}
 end
-function filter:FilterTable()
+function filter.FilterData()
     local tblFunc = {}
 
-    -- Class/Race Table
-    function tblFunc:rNew(fileName, regionName, isChecked, id)
+    -- Filter data
+    function tblFunc:getFilterList()
+        local tbl = {}
+        local nsFilter = ns.dbGlobal.filter and ns.dbGlobal.filter or {}
+
+        for k, v in pairs(nsFilter) do tbl[k] = v.name end
+
+        return tbl
+    end
+    function tblFunc:newFilter(recID, name, whoCommand, race, class)
+        local tblRacesByName, tblClassesByName = {}, {}
+
+        for _,r in pairs(ns.tblRacesByName) do -- Build Fresh Race Table
+            tblRacesByName[r.name] = {
+                fileName = r.raceFile,
+                regionName = r.name,
+                isChecked = true,
+            }
+        end
+        for _,r in pairs(ns.tblClassesByName) do -- Build Fresh Class Table
+            tblClassesByName[r.name] = {
+                color = r.color,
+                fileName = r.classFile,
+                regionName = r.name,
+                isChecked = true,
+            }
+        end
+
+        local tbl = {
+            recordID = recID or nil,
+            name = name or nil,
+            whoCommand = whoCommand or nil,
+            RACE = race or tblRacesByName,
+            CLASS = class or tblClassesByName,
+        }
+        tbl.RACE.allChecked = true
+        tbl.CLASS.allChecked = true
+
+        return tbl
+    end
+    function tblFunc:saveFilter(name, whoCommand, race, class)
         return {
-            id = id or nil,
-            fileName = fileName,
-            regionName = regionName,
-            isChecked = isChecked,
-        }
-    end
-    function tblFunc:rSave(listType, id, isChecked, fileName, regionName)
-        local tblFilter = filter.tblFilter
-        if not tblFilter then return end
-
-        local tbl = listType == 'CLASS' and tblFilter.CLASS or tblFilter.RACE
-
-        if not tbl[id] then tbl[id] = filter.filterData:rNew(fileName, regionName, isChecked)
-        else tbl[id].isChecked = isChecked end
-    end
-    function tblFunc:rGet(listType, key)
-        local tblFilter = filter.tblFilter
-        local tbl = listType == 'CLASS' and tblFilter.CLASS or tblFilter.RACE
-        if type(key) == 'string' then
-            key = strupper(key)
-            tbl = listType == 'CLASS' and tblFilter.tblClassByclassFile or tblFilter.tblRaceByclassFile end
-
-        return tbl[key]
-    end
-
-    -- Class/Race Routines
-    function tblFunc:createRecords()
-        local tblRaceByFileName = filter.tblRaceByFileName
-        local tblClassByFileName = filter.tblClassByFileName
-
-        for k, v in pairs(ns.tblRacesByName) do
-            filter.filterData:rSave('RACE', v.id, true, v.raceFile, v.name)
-            tblRaceByFileName[v.raceFile] = filter.filterData:rNew(v.raceFile, v.name, v.isChecked, k)
-        end
-        for k, v in pairs(ns.tblClassesByName) do
-            filter.filterData:rSave('CLASS', v.id, true, v.classFile, v.name)
-            tblClassByFileName[v.classFile] = filter.filterData:rNew(v.classFile, v.name, v.isChecked, k)
-        end
-    end
-
-    -- Filter Record Table
-    function tblFunc:fNew()
-        local newRecord = {
-            recordID = nil,
-            name = nil,
-            whoCommand = nil,
-            RACE = {},
-            CLASS = {},
-        }
-        newRecord.RACE.allChecked = true
-        newRecord.CLASS.allChecked = true
-
-        return newRecord
-    end
-    function tblFunc:fSave(name, whoCommand)
-        if not filter.tblFilter then return end
-
-        filter.tblFilter = {
-            recordID = filter.tblFilter.recordID or nil,
             name = name or filter.tblFilter.name,
             whoCommand = whoCommand or filter.tblFilter.whoCommand,
-            RACE = filter.tblFilter.RACE or {},
-            CLASS = filter.tblFilter.CLASS or {},
+            RACE = race or filter.tblFilter.RACE,
+            CLASS = class or filter.tblFilter.CLASS,
         }
     end
-    function tblFunc:fSaveCheck(listType, tbl)
-        if not filter.tblFilter then return end
+    function tblFunc:saveFilterData()
+        local tblBase = ns.screens.base.tblFrame
+        local nsFilter = ns.dbGlobal.filterList and ns.dbGlobal.filterList or {}
 
-        filter.tblFilter[listType] = tbl
+        -- Setup Filter Data for Saving
+        if nsFilter[filter.tblFilter.recordID] then nsFilter[filter.tblFilter.recordID] = nil end
+
+        nsFilter[filter.tblFilter.name] = {
+            name = filter.tblFilter.name,
+            whoCommand = filter.tblFilter.whoCommand,
+            RACE = filter.tblFilter.RACE,
+            CLASS = filter.tblFilter.CLASS,
+        }
+
+
+        filter.tblFilter.recordID = filter.tblFilter.name
+        tblBase.statusText:SetText(L['Filter Saved']..' '..filter.tblFilter.name)
+
+        filter.tblFrame.inline.filterCombo:SetList(filter.filterData:fGetList())
+        filter.tblFrame.inline.filterCombo:SetValue(filter.tblFilter.name)
+
+        C_Timer.After(5, function()
+            ns.screens.base.tblFrame.statusText:SetText('')
+        end)
     end
-    function tblFunc:fGetList()
-    end
+    -- Filter Race/Class data
 
     return tblFunc
 end
@@ -117,93 +116,52 @@ function filter:StartUp()
     tblBase.frame:SetSize(500, 580)
     tblBase.frame:SetShown(true)
 
-    -- Create Filter Table
-    self.tblFilter = filter.filterData:fNew()
-    self.filterData:createRecords()
+    filter.tblFilter = filter.fData:newFilter()
 
     -- Create UI
-    self.createUI:BaseFilterUI()
-    self.createUI:BuildFilterTopUI()
-    self.createUI:BuildClassListUI()
-    self.createUI:BuildRaceListUI()
+    self.cUI:BaseUI()
+    self.cUI:BuildFilterTopUI()
+    self.cUI:BuildCheckBoxListUI()
 end
+
+-- Create UI
 function filter:CreateUI()
     local tblFunc = {}
-    local skipCheckmarks = false
 
-    -- CheckBox Routines
-    local function evaluateIfAllChecked(listType)
-        local checkBoxes = filter.tblFrame.inline[listType].checkBoxes
-
-        skipCheckmarks = true
-        filter.tblFrame.inline[listType].checkBoxAll:SetValue(true)
-        for _, v in pairs(checkBoxes) do
-            if not v:GetValue() then
-                filter.tblFilter[listType].allChecked = false
-                filter.tblFrame.inline[listType].checkBoxAll:SetValue(false)
-                break
-            end
-        end
-        skipCheckmarks = false
-    end
-    local function buildCheckBoxList(listType, control)
-        local tblList = listType == 'CLASS' and ns.tblClasses or ns.tblRaces
-        local tbl = ns.code:sortTableByField(tblList, 'name')
-
-        filter.tblFrame.inline[listType].checkBoxes = {}
-        for _, v in pairs(tbl) do
-            local fName = listType == 'CLASS' and ns.code:cText(v.color, v.name) or v.name
-            local checkBox = aceGUI:Create('CheckBox')
-            checkBox:SetLabel(fName)
-            checkBox:SetValue(filter.tblFilter[listType][v.id].isChecked or false)
-            checkBox:SetRelativeWidth(.5)
-            checkBox:SetCallback('OnValueChanged', function(_, _, isChecked)
-                filter.filterData:rSave(listType, v.id, isChecked)
-                evaluateIfAllChecked(listType)
-            end)
-            control:AddChild(checkBox)
-            filter.tblFrame.inline[listType].checkBoxes[v.id] = checkBox
-        end
-    end
-    local function checkAllCheckBoxes(listType)
-        local checkBoxes = filter.tblFrame.inline[listType].checkBoxes
-        skipCheckmarks = true
-
-        for _, v in pairs(checkBoxes) do
-            v:SetValue(filter.tblFilter[listType].allChecked)
-        end
-        skipCheckmarks = false
-    end
-    local function checkControls(disableSave)
-        -- New Button
-        filter.tblFrame.inline.btnNew:SetDisabled(not (filter.tblFilter.recordID or false))
-
-        -- Delete Button
-        filter.tblFrame.inline.btnDelete:SetDisabled(not (filter.tblFilter.recordID or false))
-
-        -- Save Button
-        filter.tblFrame.inline.btnSave:SetDisabled(disableSave or (not filter.tblFilter.name))
-    end
     local function resetCustomFilter()
-        local tblFrame = filter.tblFrame
+    end
+    local function checkControls()
+    end
+    local function createCheckBoxes(scrollFrame, tbl) -- tblFilter[listType]
+        tbl.allChecked = nil
+        local tblOut = ns.code:sortTableByField(tbl, 'regionName')
 
-        filter.tblFilter = filter.filterData:fNew()
+        -- Class Checkboxes
+        for k in pairs(tblOut) do
+            local v = tblOut[k]
+            local checkbox = aceGUI:Create('CheckBox')
+            local fName = v.color and ns.code:cText(v.color, v.regionName) or v.regionName
 
-        tblFrame.inline.filterName:SetText('')
-        tblFrame.inline.whoEditbox:SetText('')
-        tblFrame.inline.filterCombo:SetList(filter.filterData:fGetList())
+            checkbox:SetLabel(fName)
+            checkbox:SetValue(v.isChecked)
+            checkbox:SetCallback('OnValueChanged', function(_, _, val)
+                v.isChecked = val
 
-        tblFrame.inline.RACE.checkBoxAll:SetValue(true)
-        tblFrame.inline.CLASS.checkBoxAll:SetValue(true)
-
-        checkAllCheckBoxes('RACE')
-        checkAllCheckBoxes('CLASS')
-
-        checkControls()
+                local allChecked = true
+                if not v.isChecked then tbl.allChecked = false
+                else
+                    for _, r in pairs(tblOut) do
+                        if not r.isChecked then allChecked = false break end
+                    end
+                    tbl.allChecked = allChecked
+                end
+            end)
+            scrollFrame:AddChild(checkbox)
+        end
     end
 
-    -- Base Filter UI
-    function tblFunc:BaseFilterUI()
+    -- Base UI
+    function tblFunc:BaseUI()
         local tblBase, tblFrame = ns.screens.base.tblFrame, filter.tblFrame
 
         local f = tblFrame.frame or CreateFrame('Frame', nil, tblBase.frame)
@@ -220,6 +178,8 @@ function filter:CreateUI()
         inline.frame:SetShown(true)
         tblFrame.inline = inline
     end
+
+    -- Build Filter Top UI
     function tblFunc:BuildFilterTopUI()
         local tblFrame = filter.tblFrame
 
@@ -228,12 +188,21 @@ function filter:CreateUI()
         inline:SetFullWidth(true)
         tblFrame.inline:AddChild(inline)
 
-        -- Title combobox
+        -- Select Filter Combobox
         local combobox = aceGUI:Create('Dropdown')
-        combobox:SetLabel(L['Custom Filters']..' (Still under development.):')
+        combobox:SetLabel(L['Custom Filters']..' (Still under development):')
         combobox:SetRelativeWidth(.5)
-        combobox:SetList(filter.filterData:fGetList())
+        combobox:SetList(filter.fData:getFilterList())
         combobox:SetCallback('OnValueChanged', function(_, _, val)
+            local nsTblFilter = ns.dbGlobal.filterList and ns.dbGlobal.filterList or {}
+
+            for _, v in pairs(nsTblFilter) do
+                if v.name == val then
+                    filter.tblFilter = filter.fData:newFilter(val, v.name, v.whoCommand, v.RACE, v.CLASS)
+                    -- Check race/class checkboxes
+                    break
+                end
+            end
         end)
         inline:AddChild(combobox)
         tblFrame.inline.filterCombo = combobox
@@ -242,9 +211,7 @@ function filter:CreateUI()
         local btnNew = aceGUI:Create('Button')
         btnNew:SetText(L['New'])
         btnNew:SetRelativeWidth(.25)
-        btnNew:SetCallback('OnClick', function()
-            resetCustomFilter()
-        end)
+        btnNew:SetCallback('OnClick', function() resetCustomFilter() end)
         inline:AddChild(btnNew)
         tblFrame.inline.btnNew = btnNew
 
@@ -253,7 +220,14 @@ function filter:CreateUI()
         btnDelete:SetText(L['Delete'])
         btnDelete:SetRelativeWidth(.25)
         btnDelete:SetCallback('OnClick', function()
+            if not filter.tblFilter.recordID then return end
 
+            local nsFilter = ns.dbGlobal.filterList and ns.dbGlobal.filterList or {}
+            nsFilter[filter.tblFilter.recordID] = nil
+
+            filter.tblFrame.inline.filterCombo:SetList(filter.filterData:fGetList())
+
+            resetCustomFilter()
         end)
         inline:AddChild(btnDelete)
         tblFrame.inline.btnDelete = btnDelete
@@ -265,18 +239,22 @@ function filter:CreateUI()
         editbox:SetCallback('OnEnterPressed', function(_,_, text)
             local name = strlower(text)
             local tblBase = ns.screens.base.tblFrame
+            local nsFilter = ns.dbGlobal.filterList and ns.dbGlobal.filterList or {}
 
             tblBase.statusText:SetText('')
-            for _, r in pairs(ns.dbGlobal.filterList and ns.dbGlobal.filterList or {}) do
-                if strlower(r.name) == name then
+            for _, r in pairs(nsFilter) do
+                if strlower(filter.tblFilter.recordID) == name then break
+                elseif strlower(r.name) == name then
+                    filter.nameOk = false
                     tblBase.statusText:SetText(L['Filter Name Already Exists'])
-                    checkControls('DISABLE_SAVE')
+                    checkControls()
                     editbox:SetFocus() -- Set focus back to the editbox
                     return
                 end
             end
 
-            filter.filterData:fSave(name)
+            filter.nameOk = true
+            filter.fData:saveFilter(name)
             checkControls()
         end)
         inline:AddChild(editbox)
@@ -286,12 +264,7 @@ function filter:CreateUI()
         local btnSave = aceGUI:Create('Button')
         btnSave:SetText(L['Save'])
         btnSave:SetRelativeWidth(.3)
-        btnSave:SetCallback('OnClick', function()
-            local tblBase = ns.screens.base.tblFrame
-
-            local tblFilters = ns.dbGlobal.filterList and ns.dbGlobal.filterList or {}
-
-        end)
+        btnSave:SetCallback('OnClick', function() filter.fData:saveFilterData() end)
         inline:AddChild(btnSave)
         tblFrame.inline.btnSave = btnSave
 
@@ -300,7 +273,7 @@ function filter:CreateUI()
         whoEditbox:SetLabel(L['Who Command'])
         whoEditbox:SetFullWidth(true)
         whoEditbox:SetCallback('OnEnterPressed', function(_, _, text)
-            filter.tblFilter.whoCommand = text
+            filter.fData:saveFilter(nil, text)
         end)
         inline:AddChild(whoEditbox)
         tblFrame.inline.whoEditbox = whoEditbox
@@ -308,76 +281,72 @@ function filter:CreateUI()
         checkControls()
     end
 
-    -- Build CheckBox List UI
-    local scrollBoxHeight = 75
-    function tblFunc:BuildClassListUI()
+    -- Build Race/Class Checkboxes UI
+    function tblFunc:BuildCheckBoxListUI() -- First so it is on the left
         local tblFrame = filter.tblFrame
-        tblFrame.inline.CLASS, tblFrame.inline.RACE = {}, {}
+        local scrollBoxHeight = 200
 
         local inline = aceGUI:Create('InlineGroup')
-        inline:SetTitle(L['Classes'])
         inline:SetLayout('Flow')
         inline:SetFullWidth(true)
         inline:SetHeight(scrollBoxHeight)
         tblFrame.inline:AddChild(inline)
-        tblFrame.inline.CLASS.inline = inline
+        tblFrame.inline.classInline = inline
 
+        -- Class Header
+        local header = aceGUI:Create('Heading')
+        header:SetText(L['Classes'])
+        header:SetRelativeWidth(.5)
+        tblFrame.inline.classInline:AddChild(header)
+
+        -- Race Header
+        header = aceGUI:Create('Heading')
+        header:SetText(L['Races'])
+        header:SetRelativeWidth(.5)
+        filter.tblFrame.inline.classInline:AddChild(header)
+
+        -- Class Select/Deselect All
         local checkBox = aceGUI:Create('CheckBox')
         checkBox:SetLabel(ns.code:cText('FFFFFF00', 'Select/Deselect All Classes'))
         checkBox:SetValue(true)
-        checkBox:SetRelativeWidth(1)
+        checkBox:SetRelativeWidth(.5)
         checkBox:SetCallback('OnValueChanged', function(_, _, isChecked)
-            if not skipCheckmarks then
-                filter.tblFilter.CLASS.allChecked = isChecked
-                checkAllCheckBoxes('CLASS')
-            end
         end)
-        inline:AddChild(checkBox)
-        tblFrame.inline.CLASS.checkBoxAll = checkBox
+        filter.tblFrame.inline.classInline:AddChild(checkBox)
+        tblFrame.inline.classCheckBoxAll = checkBox
 
-        local scrollBox = aceGUI:Create('ScrollFrame')
-        scrollBox:SetLayout('Flow')
-        scrollBox:SetFullWidth(true)
-        scrollBox:SetHeight(scrollBoxHeight)
-        inline:AddChild(scrollBox)
-
-        buildCheckBoxList('CLASS', scrollBox)
-    end
-    function tblFunc:BuildRaceListUI()
-        local tblFrame = filter.tblFrame
-
-        local inline = aceGUI:Create('InlineGroup')
-        inline:SetTitle(L['Races'])
-        inline:SetLayout('Flow')
-        inline:SetFullWidth(true)
-        inline:SetHeight(scrollBoxHeight)
-        --inline:SetPoint('TOPLEFT', tblFrame.inline.CLASS.inline.frame, 'TOPRIGHT')
-        tblFrame.inline:AddChild(inline)
-
-        local checkBox = aceGUI:Create('CheckBox')
+        -- Race Select/Deselect All
+        checkBox = aceGUI:Create('CheckBox')
         checkBox:SetLabel(ns.code:cText('FFFFFF00', 'Select/Deselect All Races'))
         checkBox:SetValue(true)
-        checkBox:SetRelativeWidth(1)
+        checkBox:SetRelativeWidth(.5)
         checkBox:SetCallback('OnValueChanged', function(_, _, isChecked)
-            if not skipCheckmarks then
-                filter.tblFilter.RACE.allChecked = isChecked
-                checkAllCheckBoxes('RACE')
-            end
         end)
-        inline:AddChild(checkBox)
-        tblFrame.inline.RACE.checkBoxAll = checkBox
+        filter.tblFrame.inline.classInline:AddChild(checkBox)
+        tblFrame.inline.raceCheckBoxAll = checkBox
 
-        local scrollBox = aceGUI:Create('ScrollFrame')
-        scrollBox:SetLayout('Flow')
-        scrollBox:SetFullWidth(true)
-        scrollBox:SetHeight(scrollBoxHeight)
-        inline:AddChild(scrollBox)
+        -- Class Scroll Frame
+        local scrollFrame = aceGUI:Create('ScrollFrame')
+        scrollFrame:SetLayout('Flow')
+        scrollFrame:SetRelativeWidth(.5)
+        scrollFrame:SetHeight(scrollBoxHeight)
+        inline:AddChild(scrollFrame)
 
-        buildCheckBoxList('RACE', scrollBox)--]]
+        createCheckBoxes(scrollFrame, filter.tblFilter.CLASS)
+
+        -- Race Scroll Frame
+        scrollFrame = aceGUI:Create('ScrollFrame')
+        scrollFrame:SetLayout('Flow')
+        scrollFrame:SetRelativeWidth(.5)
+        scrollFrame:SetHeight(scrollBoxHeight)
+        filter.tblFrame.inline.classInline:AddChild(scrollFrame)
+
+        createCheckBoxes(scrollFrame, filter.tblFilter.RACE)
     end
 
     return tblFunc
 end
-filter.createUI = filter:CreateUI()
-filter.filterData = filter:FilterTable()
 filter:Init()
+
+filter.cUI = filter:CreateUI()
+filter.fData = filter:FilterData()